+++ src/main/java/com/parking/management/module/incident/IncidentReportRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface IncidentReportRepository extends JpaRepository<IncidentReport, Long> {
+public interface IncidentReportRepository extends JpaRepository<IncidentReport, Integer> {
     // Custom query 1
     // List<IncidentReport> findBySomeField(String field);
 }
